+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -104,7 +104,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdwqh4etxkrcb4t64wrwq5da"
+        "conv_01kdx3cm323m2rc9rmhf2qjc6y"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """