+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDWQH4XB4NA9GF1KV09E5J88"
+        "01KDX3CMG58H0FAP0PX9PRECH6"
     )
     r"""The unique identifier of the data source"""