+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveconversationop.py
@@ -112,7 +112,7 @@ class RetrieveConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdwqh4ezn99bj6wzwz9sq0t3"
+        "conv_01kdx3cm34sf30h5nb8aqnyzw3"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """