+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -80,5 +80,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-01T12:11:30.915Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-01T15:38:46.209Z")
     r"""The date and time the resource was last updated"""