+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDWQH4XE6W8C2QBYXFB43FFV"
+        "01KDX3CMG6EMGSJKEP1GAF7BDT"
     )
     r"""The unique identifier of the data source"""